+++ src/components/Post/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
-import "./Post.css";
-import { db } from "./firebase";
+import "./style.css";
+import { db } from "../../firebase";
 import firebase from "firebase";
 import Avatar from "@material-ui/core/Avatar";
